UserModel created
Classes are separated into models and resources folders
