+++ section_6_sqlalchemy/code/resources/user.py
@@ -1,50 +1,6 @@
 import sqlite3
 from flask_restful import Resource, reqparse
-
-class User:
-    def __init__(self, _id, username, password):
-        self.id = _id
-        self.username = username
-        self.password = password
-
-    @classmethod
-    def find_by_username(cls, username):
-        connection = sqlite3.connect('data.db')
-        cursor = connection.cursor()
-
-        query = 'SELECT * FROM users WHERE username = ?'
-        result = cursor.execute(query, (username,))
-        #get the first row
-        row = result.fetchone()
-        #if row is not None
-        if row:
-            #user = cls(row[0], row[1], row[2])
-            user = cls(*row)
-        else:
-            user = None
-
-        connection.close()
-        return user
-
-    @classmethod
-    def find_by_id(cls, _id):
-        connection = sqlite3.connect('data.db')
-        cursor = connection.cursor()
-
-        query = 'SELECT * FROM users WHERE id = ?'
-        result = cursor.execute(query, (_id,))
-        #get the first row
-        row = result.fetchone()
-        #if row is not None
-        if row:
-            #user = cls(row[0], row[1], row[2])
-            user = cls(*row)
-        else:
-            user = None
-
-        connection.close()
-        return user
-
+from models.user import UserModel
 
 class UserRegister(Resource):
     parser = reqparse.RequestParser()
@@ -66,7 +22,7 @@ class UserRegister(Resource):
 
         data = UserRegister.parser.parse_args()
 
-        if User.find_by_username(data['username']):
+        if UserModel.find_by_username(data['username']):
             return {"message":"A user with that username already exists!"}
 
         connection = sqlite3.connect('data.db')
